+++ rust/operator-binary/src/main.rs
@@ -108,7 +108,7 @@ async fn main() -> anyhow::Result<()> {
 
             let (ctx, cm_store_tx, secret_store_tx) = create_context(client.clone());
 
-            let (webhook_server, initial_reconcile_rx) = create_webhook_server(
+            let webhook_server = create_webhook_server(
                 ctx.clone(),
                 &operator_environment,
                 disable_restarter_mutating_webhook,
@@ -117,38 +117,26 @@ async fn main() -> anyhow::Result<()> {
             )
             .await?;
 
-            // Multiply initial reconcile signal for the STS and pod restart controllers.
-            let initial_reconcile_signal = SignalWatcher::new(initial_reconcile_rx.map(|_| ()));
-
-            let delayed_sts_restart_controller = async {
-                initial_reconcile_signal.handle().await;
-
-                restart_controller::statefulset::start(
-                    ctx,
-                    cm_store_tx,
-                    secret_store_tx,
-                    &watch_namespace,
-                    sigterm_watcher.handle(),
-                )
-                .await
-            }
+            let sts_restart_controller = restart_controller::statefulset::start(
+                ctx,
+                cm_store_tx,
+                secret_store_tx,
+                &watch_namespace,
+                sigterm_watcher.handle(),
+            )
             .map(anyhow::Ok);
 
-            let delayed_pod_restart_controller = async {
-                initial_reconcile_signal.handle().await;
-
+            let pod_restart_controller =
                 restart_controller::pod::start(&client, &watch_namespace, sigterm_watcher.handle())
-                    .await
-            }
-            .map(anyhow::Ok);
+                    .map(anyhow::Ok);
 
             let webhook_server = webhook_server
                 .run(sigterm_watcher.handle())
                 .map_err(|err| anyhow!(err).context("failed to run webhook"));
 
             futures::try_join!(
-                delayed_sts_restart_controller,
-                delayed_pod_restart_controller,
+                sts_restart_controller,
+                pod_restart_controller,
                 webhook_server,
                 eos_checker,
             )?;